이제부터는 양의 정수 뿐만 아니라 실수, 즉 double 타입의 값을 전달 받았을 경우에도 연산이 수행되도록, 즉 피연산자를 여러 타입으로 받을 수 있도록 기능을 확장하고 싶습니다.

+++ src/main/java/calculator/App.java
@@ -29,8 +29,8 @@ public class App {
             }
 
             if (calculateIntent == 1) {
-                int num1;
-                int num2;
+                double num1;
+                double num2;
 
                 try {
                     System.out.print("첫 번째 숫자를 입력하세요: ");
